20231108-product-008 정수와 정수끼리의 연산은 나눗셈도 정수입니다.

+++ src/main/java/com/mergeco/oiljang/product/controller/ProductController.java
@@ -70,7 +70,7 @@ public class ProductController {
         int offset = limit * (page - 1);
         List<ProductListDTO> productListDTOList = productService.selectProductList(offset, limit, categoryCode, sortCondition, minPrice, maxPrice);
 
-        int totalItem = Long.valueOf(productService.countProductList()).intValue();
+        double totalItem = Long.valueOf(productService.countProductList()).doubleValue();
         int totalPage = (int) Math.ceil(totalItem / limit);
 
         if(page >= totalPage) {
